refactor(front): split unit tests into domain files per writing-unit-tests skill
